pyproject.toml
[project]
name = "bonsai-sensei"
version = "0.1.0"
description = "FastAPI application with uv"
requires-python = ">=3.12"
dependencies = [
    "deep-translator>=1.11.4",
    "fastapi",
    "google-adk>=1.22.1",
    "litellm>=1.65.5",
    "monocle_apptrace>=0.7.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "sqlmodel>=0.0.31",
    "tavily-python>=0.7.19",
    "uvicorn",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bonsai_sensei"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
    "pyhamcrest>=2.1.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (select with '-m integration')",
]
python_functions = [
    "should_*",
]
